create new 'execute' phase for automated code execution in dynamic analysis worker package
